Update analysis with final results
This includes a change to parsing.py to move a field from float to a
string since a later version of HTML included a number large enough
to include a comma.  A note has been made that it should be updated
with a different handler going forward.

Other updates are to input files, output files, and an event summary
in README.md.